+++ src/robot_impl.cpp
@@ -2,8 +2,6 @@
 // Use of this source code is governed by the Apache-2.0 license, see LICENSE
 #include "robot_impl.h"
 
-#include <sstream>
-
 // `using std::string_literals::operator""s` produces a GCC warning that cannot be disabled, so we
 // have to use `using namespace ...`.
 // See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=65923#c0
@@ -13,24 +11,25 @@ namespace franka {
 
 namespace {
 inline ControlException createControlException(const CommandException& command_exception,
-                                               const RobotState& robot_state) {
+                                               const RobotState& robot_state,
+                                               const std::vector<Record>& log = {}) {
   if (robot_state.robot_mode == RobotMode::kReflex) {
-    return ControlException(command_exception.what() + " "s +
-                            static_cast<std::string>(robot_state.last_motion_errors));
+    return ControlException(
+        command_exception.what() + " "s + static_cast<std::string>(robot_state.last_motion_errors),
+        log);
   }
-  return ControlException(command_exception.what());
+  return ControlException(command_exception.what(), log);
 }
 }  // anonymous namespace
 
-Robot::Impl::Impl(std::unique_ptr<Network> network, RealtimeConfig realtime_config)
-    : network_{std::move(network)}, realtime_config_{realtime_config} {
+Robot::Impl::Impl(std::unique_ptr<Network> network, size_t log_size, RealtimeConfig realtime_config)
+    : network_{std::move(network)}, logger_{log_size}, realtime_config_{realtime_config} {
   if (!network_) {
     throw std::invalid_argument("libfranka robot: Invalid argument");
   }
 
   connect<research_interface::robot::Connect, research_interface::robot::kVersion>(*network_,
                                                                                    &ri_version_);
-
   updateState(network_->udpBlockingReceive<research_interface::robot::RobotState>());
 }
 
@@ -39,9 +38,13 @@ RobotState Robot::Impl::update(
     const research_interface::robot::ControllerCommand* control_command) {
   network_->tcpThrowIfConnectionClosed();
 
-  sendRobotCommand(motion_command, control_command);
+  research_interface::robot::RobotCommand robot_command =
+      sendRobotCommand(motion_command, control_command);
 
-  return convertRobotState(receiveRobotState());
+  RobotState state = convertRobotState(receiveRobotState());
+  logger_.log(state, robot_command);
+
+  return state;
 }
 
 void Robot::Impl::throwOnMotionError(const RobotState& robot_state, uint32_t motion_id) {
@@ -57,7 +60,7 @@ void Robot::Impl::throwOnMotionError(const RobotState& robot_state, uint32_t mot
       handleCommandResponse<research_interface::robot::Move>(
           network_->tcpBlockingReceiveResponse<research_interface::robot::Move>(motion_id));
     } catch (const CommandException& e) {
-      throw createControlException(e, robot_state);
+      throw createControlException(e, robot_state, logger_.flush());
     }
   }
 }
@@ -71,11 +74,11 @@ RobotState Robot::Impl::readOnce() {
   return convertRobotState(receiveRobotState());
 }
 
-void Robot::Impl::sendRobotCommand(
+research_interface::robot::RobotCommand Robot::Impl::sendRobotCommand(
     const research_interface::robot::MotionGeneratorCommand* motion_command,
     const research_interface::robot::ControllerCommand* control_command) const {
+  research_interface::robot::RobotCommand robot_command{};
   if (motion_command != nullptr || control_command != nullptr) {
-    research_interface::robot::RobotCommand robot_command{};
     robot_command.message_id = message_id_;
     if (motion_command != nullptr) {
       if (!motionGeneratorRunning()) {
@@ -100,6 +103,7 @@ void Robot::Impl::sendRobotCommand(
 
     network_->udpSend<research_interface::robot::RobotCommand>(robot_command);
   }
+  return robot_command;
 }
 
 research_interface::robot::RobotState Robot::Impl::receiveRobotState() {
@@ -204,11 +208,13 @@ uint32_t Robot::Impl::startMotion(
         break;
       }
     } catch (const CommandException& e) {
-      throw createControlException(e, robot_state);
+      throw createControlException(e, robot_state, logger_.flush());
     }
 
     robot_state = update();
   }
+
+  logger_.flush();
 
   return move_command_id;
 }