+++ test/helpers.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "franka/gripper_state.h"
+#include "franka/log.h"
 #include "franka/robot_state.h"
 #include "research_interface/gripper/types.h"
 #include "research_interface/robot/rbk_types.h"
@@ -23,6 +24,10 @@ void testMotionGeneratorCommandsAreEqual(
     const research_interface::robot::MotionGeneratorCommand& actual);
 void testControllerCommandsAreEqual(const research_interface::robot::ControllerCommand& expected,
                                     const research_interface::robot::ControllerCommand& actual);
+void testRobotCommandsAreEqual(const research_interface::robot::RobotCommand& expected,
+                               const research_interface::robot::RobotCommand& actual);
+void testRobotCommandsAreEqual(const research_interface::robot::RobotCommand& expected,
+                               const franka::RobotCommand actual);
 
 namespace research_interface {
 namespace robot {